Create README.md
This repository is a collection of prompts developed by farahcube17 as part of my journey to becoming a prompt engineer.  It showcases a variety of prompts created for different purposes, including natural language processing, giving advice, etc. This portfolio aims to demonstrate the ability to craft high-quality prompts in the Indonesian Language that effectively utilize AI models such as Bard AI, Open AI, etc.